requirements.txt
nltk
pattern
pyenchant
